back/server/requirements.txt
alembic==1.6.5
asgiref==3.5.1
backports.entry-points-selectable==1.1.1
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.1
colorama==0.4.4
config==0.5.1
cryptography==37.0.2
defusedxml==0.7.1
distlib==0.3.3
Django==4.0.4
django-allauth==0.50.0
django-cors-headers==3.12.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
dnspython==2.1.0
email-validator==1.1.3
filelock==3.4.0
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
jsonparser==1.0
Mako==1.1.4
MarkupSafe==2.0.1
migrate==0.3.7
mysql-connector-python==8.0.26
numpy==1.22.1
oauthlib==3.2.0
platformdirs==2.4.0
protobuf==3.17.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-editor==1.0.4
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.22
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
virtualenv-clone==0.5.7
